+++ src/models/Plano.h
@@ -7,10 +7,10 @@ namespace Modelos
     class Plano : public Modelo
 {
 public:
-    int tamanho;
+    float tamanho;
     float subDivisions;
 
-    Plano(int l,float sub);
+    Plano(float l,float sub);
     void saveModel(std::ofstream &file);
 private:
     void savePlaneTXZ(float y,std::ofstream &file);
